requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coverage==5.5
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.2
Markdown==3.3.4
openpyxl==3.0.7
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
